Integrated SKU based Product and Mockup Creation

+++ src/main/java/com/example/madrasdaapi/controllers/AdminController.java
@@ -3,7 +3,9 @@ package com.example.madrasdaapi.controllers;
 import com.example.madrasdaapi.dto.AuthDTO.RegisterDTO;
 import com.example.madrasdaapi.dto.VendorDTO.VendorDTO;
 import com.example.madrasdaapi.dto.VendorDTO.VendorMenuItemDTO;
+import com.example.madrasdaapi.exception.ResourceNotFoundException;
 import com.example.madrasdaapi.repositories.UserRepository;
+import com.example.madrasdaapi.repositories.VendorRepository;
 import com.example.madrasdaapi.services.AdminServices.AdminService;
 import com.example.madrasdaapi.services.AdminServices.MockupService;
 import com.example.madrasdaapi.services.VendorServices.VendorService;
@@ -23,6 +25,7 @@ public class AdminController {
      private final AdminService adminService;
      private final MockupService mockupService;
      private final UserRepository userRepository;
+     private final VendorRepository vendorRepository;
 
 //     @PreAuthorize("hasRole('ADMIN')")
      @GetMapping("getVendors")
@@ -41,7 +44,7 @@ public class AdminController {
 //     @PreAuthorize("hasRole('ADMIN')")
      @DeleteMapping("deleteVendor/{id}")
      public void deleteVendor(@PathVariable Long id) {
-
+          if(!vendorRepository.existsById(id)) throw new ResourceNotFoundException("Vendor", "id", id);
           adminService.deleteVendor(id);
      }
